+++ crates/network-libp2p/src/peers/score.rs
@@ -12,13 +12,12 @@ use std::time::Instant;
 use tn_config::ScoreConfig;
 
 /// Global static configuration that is initialized only once for all peers.
-pub(super) static GLOBAL_SCORE_CONFIG: OnceLock<Arc<ScoreConfig>> = OnceLock::new();
+pub(crate) static GLOBAL_SCORE_CONFIG: OnceLock<Arc<ScoreConfig>> = OnceLock::new();
 
 /// Initialize the global peer score configuration.
-#[must_use]
 pub(super) fn init_peer_score_config(config: ScoreConfig) {
     let config = Arc::new(config);
-    GLOBAL_SCORE_CONFIG.set(config).expect("Peer score configuration initialized once")
+    GLOBAL_SCORE_CONFIG.set(config).expect("Peer score configuration initialized once");
 }
 
 /// Get a reference to the global peer score configuration
@@ -243,6 +242,7 @@ impl Display for Score {
 
 /// The expected status of the peer based on the peer's score.
 #[derive(Debug, PartialEq, Clone, Copy)]
+// TODO: this was ScoreState
 pub(super) enum Reputation {
     /// The peer is performing within the tolerable threshold.
     Trusted,
